[123] Fixed auctionshouse / auctions opcodes Fixed tickets. Fixed some chat opcodes. Fixed fishing opcodes. Fixed commentator opcodes. Fixed...
Added 3 new opcodes:
- CMSG_REQUEST_RATED_BATTLEGROUND_INFO.
- CMSG_REQUEST_PVP_OPTIONS_ENABLED.
- CMSG_COMMENTATOR_START_WARGAME.#


Fixed ace debug build.

+++ Core/src/main/Achievements/AchievementMgr.cpp
@@ -425,7 +425,7 @@ void AchievementMgr::Reset()
     CheckAllAchievementCriteria();
 }
 
-void AchievementMgr::ResetAchievementCriteria(AchievementCriteriaTypes type, uint32 miscvalue1, uint32 miscvalue2, bool evenIfCriteriaComplete)
+void AchievementMgr::ResetAchievementCriteria(AchievementCriteriaTypes type, uint64 miscvalue1, uint64 miscvalue2, bool evenIfCriteriaComplete)
 {
     sLog->outDebug(LOG_FILTER_ACHIEVEMENTSYS, "AchievementMgr::ResetAchievementCriteria(%u, %u, %u)", type, miscvalue1, miscvalue2);
 
@@ -729,7 +729,7 @@ static const uint32 achievIdForDangeon[][4] =
 /**
  * this function will be called whenever the user might have done a criteria relevant action
  */
-void AchievementMgr::UpdateAchievementCriteria(AchievementCriteriaTypes type, uint32 miscvalue1, uint32 miscvalue2, Unit *unit, uint32 time)
+void AchievementMgr::UpdateAchievementCriteria(AchievementCriteriaTypes type, uint64 miscvalue1, uint64 miscvalue2, Unit *unit, uint32 time)
 {
     sLog->outDebug(LOG_FILTER_ACHIEVEMENTSYS, "AchievementMgr::UpdateAchievementCriteria(%u, %u, %u, %u)", type, miscvalue1, miscvalue2, time);
